BitsetView: use long int to make popcount faster (callgrind ir/call 2227 -> 1135)

+++ src/search/per_state_bitset.h
@@ -9,7 +9,7 @@
 
 class BitsetMath {
 public:
-    using Block = unsigned int;
+    using Block = unsigned long int;
     static_assert(
         !std::numeric_limits<Block>::is_signed,
         "Block type must be unsigned");
